refactor: move CryptoData to repository layer
