+++ app/src/main/java/com/amsterdam/aflami/di/repositoryModule.kt
@@ -12,6 +12,8 @@ import com.example.repository.repository.CountryRepositoryImpl
 import com.example.repository.repository.MovieRepositoryImpl
 import com.example.repository.repository.RecentSearchRepositoryImpl
 import com.example.repository.repository.TvShowRepositoryImpl
+import com.example.repository.security.CryptoData
+import com.example.repository.security.CryptoDataImpl
 import com.example.repository.utils.RecentSearchHandler
 import com.example.repository.utils.RecentSearchHandlerImpl
 import org.koin.core.module.dsl.singleOf
@@ -19,6 +21,8 @@ import org.koin.dsl.bind
 import org.koin.dsl.module
 
 val repositoryModule = module {
+    // crypto
+    singleOf(::CryptoDataImpl) bind CryptoData::class
 
     // Handler
     singleOf(::RecentSearchHandlerImpl) bind RecentSearchHandler::class
