+++ localDatasource/src/main/java/com/example/localdatasource/dataStore/appPreferences/AppDataStorePreferences.kt
@@ -4,16 +4,15 @@ import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
+import com.example.domain.exceptions.UnauthorizedException
 import com.example.localdatasource.dataStore.appPreferences.AppDataStorePreferences.PreferenceKeys.SESSION_ID
 import com.example.localdatasource.dataStore.appPreferences.AppDataStorePreferences.PreferenceKeys.SESSION_TYPE
 import com.example.localdatasource.dataStore.datasource.AppPreferences
-import com.example.localdatasource.dataStore.utils.CryptoData
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 
 class AppDataStorePreferences(
     private val datastore: DataStore<Preferences>,
-    val cryptoData: CryptoData,
 ) : AppPreferences {
     private object PreferenceKeys {
         val SESSION_TYPE = stringPreferencesKey("sessionType")
@@ -33,19 +32,16 @@ class AppDataStorePreferences(
     }
 
     override suspend fun cacheSessionId(sessionId: String) {
-        val encryptedSessionId = cryptoData.encryptString(sessionId)
         datastore.edit { setting ->
-            setting[SESSION_ID] = encryptedSessionId
+            setting[SESSION_ID] = sessionId
         }
     }
 
     override suspend fun getSessionId(): String? {
-        val encryptedSessionId =
-            datastore.data
-                .map { setting ->
-                    setting[SESSION_ID] ?: throw Exception()
-                }.first()
-        return cryptoData.decryptString(encryptedSessionId)
+        return datastore.data
+            .map { setting ->
+                setting[SESSION_ID] ?: throw UnauthorizedException()
+            }.first()
     }
 
     override suspend fun clearSessionId() {
