--- localDatasource/src/main/java/com/example/localdatasource/dataStore/utils/CryptoData.kt
@@ -1,7 +1,0 @@
-package com.example.localdatasource.dataStore.utils
-
-interface CryptoData {
-    fun encryptString(input: String): String
-
-    fun decryptString(input: String): String?
-}
