+++ repository/src/main/java/com/example/repository/repository/AuthenticationRepositoryImpl.kt
@@ -1,34 +1,38 @@
 package com.example.repository.repository
 
+import com.example.domain.exceptions.UnknownException
 import com.example.domain.repository.AuthenticationRepository
 import com.example.domain.utils.SessionType
 import com.example.repository.datasource.local.AuthenticationLocalSource
 import com.example.repository.datasource.remote.AuthenticationRemoteSource
 import com.example.repository.mapper.local.SessionTypeMapper
+import com.example.repository.security.CryptoData
 
 class AuthenticationRepositoryImpl(
     private val authenticationRemoteSource: AuthenticationRemoteSource,
     private val authenticationLocalSource: AuthenticationLocalSource,
-    private val sessionTypeMapper: SessionTypeMapper
+    private val sessionTypeMapper: SessionTypeMapper,
+    val cryptoData: CryptoData,
 ) : AuthenticationRepository {
     override suspend fun loginWithPassword(
         username: String,
         password: String,
     ) {
-        val sessionId = authenticationRemoteSource.loginWithPassword(username, password)
-        authenticationLocalSource.cacheSessionId(sessionId)
+        authenticationRemoteSource.loginWithPassword(username, password).let { sessionId ->
+            authenticationLocalSource.cacheSessionId(cryptoData.encryptString(sessionId))
+        }
         authenticationLocalSource.setSessionType(sessionTypeMapper.toLocalSessionType(SessionType.LOGGED_IN))
     }
 
-    override suspend fun getSessionId(): String = authenticationLocalSource.getCachedSessionId()
+    override suspend fun getSessionId(): String =
+        cryptoData.decryptString(authenticationLocalSource.getCachedSessionId()) ?: throw UnknownException()
 
     override suspend fun setSessionType(sessionType: SessionType) {
         authenticationLocalSource.setSessionType(sessionTypeMapper.toLocalSessionType(sessionType))
     }
 
-    override suspend fun getSessionType(): SessionType {
-        return sessionTypeMapper.fromLocalSessionType(authenticationLocalSource.getSessionType())
-    }
+    override suspend fun getSessionType(): SessionType =
+        sessionTypeMapper.fromLocalSessionType(authenticationLocalSource.getSessionType())
 
     override suspend fun logout() {
         authenticationLocalSource.clearCachedSessionId()
