+++ repository/src/main/java/com/example/repository/security/CryptoData.kt
@@ -1,0 +1,11 @@
+package com.example.repository.security
+
+interface CryptoData {
+    fun encrypt(bytes: ByteArray): ByteArray
+
+    fun decrypt(bytes: ByteArray): ByteArray?
+
+    fun encryptString(input: String): String
+
+    fun decryptString(input: String): String?
+}
